Create internal knowledge base server
Create internal knowledge base server used to handle knowledge
processing (storage and revival). The server is implemented using Flask
and the SQLite database; however, the database engine is subject to
change in the future.

Currently, the implementation of the knowledge server limits to the
database portion for ConceptNet knowledge. The ConceptNet and WordNet
API will be implemented later.