Reduced to English subset of ConceptNet
After experimentation, it is found that the entirety of ConceptNet is
too large to be process by this implementation. Since language other
than English benefits very little to the project, a subset of ConceptNet
is processed and used instead.

EnglishNet - a subset of ConceptNet where every concept is an English
concept (concept with uri of the form "/c/en/.+")